messing around wit restructure
