+++ tactical_server/routes.py
@@ -1,9 +1,8 @@
-####################### Imports######################################
 import datetime  # gets date and time from computer.
 import json  # used to put data into json format.
 import os  # package used to get file path.
 
-from flask import Flask, jsonify,  redirect, render_template, url_for
+from flask import Flask, jsonify,  redirect, render_template, url_for, Blueprint
 from flask_migrate import Migrate  # used to manage changes to database.
 from flask_sqlalchemy import SQLAlchemy  # package used to handle data base.
 from geopy import distance
@@ -15,272 +14,13 @@ from forms_section.forms_section import (  # forms to take input from user.
 from lat_long_to_grid_reference.lat_long_to_grid_reference import \
     lat_long_to_grid_reference  # converts lat and long to grid reference.
 
-app = Flask(__name__)  # creates flask app.
 
-app.config['SECRET_KEY'] = 'mySecretKey'  # used for encryption.
+from .extensions import db
+from .models import Soldier, Vitals, Location, Section
 
+main = Blueprint("main", __name__)
 
-########## Make SQL File ##################
-
-# gets path to current file.
-basdir = os.path.abspath(os.path.dirname(__file__))
-# creates sqlite database in same folder.
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + \
-    os.path.join(basdir, 'section_db.sqlite')
-# turn off modification tracking.
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-db = SQLAlchemy(app)  # create Sqlalchemy app.
-Migrate(app, db)  # create migration object.
-
-######### Tables #########################
-
-
-class Section(db.Model):
-    """
-    Represents a section within the system, tracking various attributes and states.
-
-    This class corresponds to the 'Section' table in the database.
-
-    Attributes:
-        id (int): Identifying number of the section, serving as the primary key.
-        section_amunition (int): Total ammunition available in the section.
-        section_strength (int): Number of personnel in the section.
-        section_ok (int): Number of uninjured soldiers in the section.
-        section_casualty (int): Number of soldiers in the section who have become casualties.
-        section_location (str): Location of the section, obtained from the section commander.
-        section_battery (str): Battery level indication for the section database.
-        Soldier (relationship): One-to-many relationship to the 'Soldier' table.
-
-    Methods:
-        __init__(self, id, section_ammo):
-            Initializes a new Section instance.
-
-            Args:
-                id (int): Identifying number of the section, must be unique.
-                section_ammo (int): Starting ammunition for the section.
-    """
-
-    __tablename__ = 'sections'  # name of the table.
-
-    # section number used as primary key.
-    id = db.Column(db.Integer, primary_key=True)
-    # total ammunition in the section.
-    section_amunition = db.Column(db.Integer)
-    # Number of people in the section.
-    section_strength = db.Column(db.Integer)
-    # Soldiers in section who are uninjured.
-    section_ok = db.Column(db.Integer)
-    # Soldiers in section who have become casualties.
-    section_casualty = db.Column(db.Integer)
-    # Location of section, Taken from section commanders location.
-    section_location = db.Column(db.Text)
-    # battery level indication for Section database.
-    section_battery = db.Column(db.Text)
-    # relationship to table soldiers.
-    soldiers = db.relationship('Soldier', backref='sections', lazy='dynamic')
-
-    def __init__(self, id, section_ammo):
-        """
-        Initializes a new section.
-
-        To create the section, it requires an entry of an id number and ammunition.
-
-        Args:
-            id (int): Identifying number of the section, must be unique.
-            section_ammo (int): Sections starting ammunition.
-        """
-        self.id = id  # Identifying number of the section must be unique.
-        self.section_amunition = section_ammo  # Sections starting ammunition.
-
-
-class Soldier(db.Model):
-    """
-    Represents an individual soldier within the system, providing detailed information about each soldier.
-
-    This class corresponds to the 'soldiers' table in the database.
-
-    Attributes:
-        id (int): Army number of the soldier, serving as the primary key. Must be unique.
-        name (str): Soldier's name.
-        role (str): Role of the section member.
-        section_id (int): ID number of the section that the soldier belongs to.
-        identity_check (str): Used to verify the soldier's identity.
-        vitals (relationship): One-to-many relationship to the 'vitals' table, listing the soldier's heart rate over time.
-        location_history (relationship): One-to-many relationship to the 'location' table, listing the soldier's location over time.
-        current_location (str): Stores the current location of the soldier.
-        last_heart_rate (int): Stores the most recent record of the soldier's heart rate.
-        last_update_time (str): Stores the time of the last update.
-        ammunition_expended (int): Stores the number of rounds fired by the soldier.
-        rifle_sensor_battery_level (str): Stores the battery level of the ammunition tracker.
-        status (str): Stores the status of the soldier.
-        armour_sensor_battery_level (str): Stores the battery level of the body armour sensor.
-        hub_sensor_battery_level (str): Stores the battery level of the sensor hub.
-        distance_traveled (str): Stores the distance traveled by the soldier.
-
-    Methods:
-        __init__(self, id, name, role, section_id):
-            Initializes a new Soldier instance.
-
-            Args:
-                id (int): Army number, must be unique.
-                name (str): Soldier's name.
-                role (str): Soldier's role in the section.
-                section_id (int): Section ID to which the soldier belongs.
-    """
-
-    __tablename__ = 'soldiers'  # Table name.
-
-    # Army number  of soldier primary key. Must be unique.
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.Text)  # Soldiers name.
-    role = db.Column(db.Text)  # Role of section member.
-    # ID number of section that soldier belongs to.
-    section_id = db.Column(db.Integer, db.ForeignKey('sections.id'))
-    identity_check = db.Column(db.Text)  # USed to verify soldiers identity.
-    # Link  to table that  lists soldiers heart rate over time.
-    vitals_history = db.relationship(
-        'Vitals', backref='soldier', lazy='dynamic')
-    # Link to table that lists soldiers location over times.
-    location_history = db.relationship(
-        'Location', backref='soldier', lazy='dynamic')
-    # Stores current location of soldier.
-    current_location = db.Column(db.Text)
-    # Stores most recent record of soldiers location.
-    last_heart_rate = db.Column(db.Integer)
-    last_update_time = db.Column(db.Text)  # Stores time of last update.
-    # Stores number of rounds fired by soldier.
-    ammunition_expended = db.Column(db.Integer)
-    # Stores battery level of Ammunition tracker.
-    rifle_sensor_battery_level = db.Column(db.Text)
-    status = db.Column(db.Text)  # Stores Status of soldier.
-    # Stores battery level of body armour sensor.
-    armour_sensor_battery_level = db.Column(db.Text)
-    # Battery level of sensor hub.
-    hub_sensor_battery_level = db.Column(db.Text)
-    # Stores distance traveled by soldier.
-    distance_traveled = db.Column(db.Text)
-
-    def __init__(self, id, name, role, section_id):
-        """
-        Initializes a new section member.
-
-        To add a soldier, it requires an entry of an army number, a name,
-        the role in the section, and the section they belong to.
-
-        Args:
-            id (int): Army number, must be unique.
-            name (str): Soldier's name.
-            role (str): Soldier's role in the section.
-            section_id (int): Section the soldier belongs to.
-        """
-        self.id = id  # Army number.
-        self.name = name  # Soldiers name.
-        self.role = role  # Soldiers role in the section.
-        self.section_id = section_id  # Section soldier belongs to.
-
-
-class Vitals(db.Model):
-    """
-    Represents a table for recording a soldier's vital signs, particularly heart rate.
-
-    This class corresponds to the 'vitals' table in the database.
-
-    Attributes:
-        id (int): Automatically generated unique primary key for the database entry.
-        heart_rate (int): Heart rate of the soldier.
-        update_time (str): Time when the heart rate entry was made.
-        soldier_id (int): ID number of the soldier to whom the entry belongs.
-
-    Methods:
-        __init__(self, heart_rate, update_time, soldier_id):
-            Initializes a new entry for heart rate for a soldier.
-
-    """
-
-    __tablename__ = 'vitals'  # Name of table.
-
-    # ID number of entry in database. automatically generated unique primary key.
-    id = db.Column(db.Integer, primary_key=True)
-    heart_rate = db.Column(db.Integer)  # Heart rate of soldier.
-    update_time = db.Column(db.Text)  # Time that heart rate entry was made.
-    # ID number of Soldier.
-    soldier_id = db.Column(db.Integer, db.ForeignKey('soldiers.id'))
-    # soldier = db.relationship('Soldier', backref='vitals_history', lazy='dynamic')
-
-    # function used to create a new entry for heart rate for soldier.Automatically created from data sent from sensors.
-    def __init__(self, heart_rate, time, soldier_id):
-        """
-        Initializes a new entry for heart rate for a soldier.
-
-        Parameters:
-            heart_rate (int): Heart rate.
-            update_time (str): Time of the entry.
-            soldier_id (int): ID number of the soldier that the entry belongs to.
-        """
-
-        self.heart_rate = heart_rate  # Heart rate.
-        self.update_time = time  # Time of entry.
-        # ID number of soldier that the entry belongs to.
-        self.soldier_id = soldier_id
-
-
-class Location(db.Model):
-    """
-    Represents a table for recording the locations a soldier has been.
-
-    This class corresponds to the 'locations' table in the database.
-
-    Attributes:
-        id (int): Automatically generated unique primary key for the database entry.
-        long (str): Longitude of the soldier's location.
-        lat (str): Latitude of the soldier's location.
-        grid_reference (str): Grid reference of the soldier's location.
-        update_time (str): Time when the location entry was made.
-        soldier_id (int): ID number of the soldier to whom the entry belongs.
-
-    Methods:
-        __init__(self, long, lat, grid_reference, update_time, soldier_id):
-            Initializes a new entry for the location of a soldier.
-
-    """
-
-    __tablename__ = 'locations'  # Table name.
-
-    # ID number of entry in database. automatically generated unique primary key.
-    id = db.Column(db.Integer, primary_key=True)
-    long = db.Column(db.Text)  # longitude.
-    lat = db.Column(db.Text)  # latitude.
-    grid_reference = db.Column(db.Text)  # Grid reference
-    update_time = db.Column(db.Text)  # Time that location entry was made.
-    # ID number of soldier that the entry belongs to.
-    soldier_id = db.Column(db.Integer, db.ForeignKey('soldiers.id'))
-    # soldier = db.relationship('Soldier', backref='location_history', lazy='dynamic')
-
-    # function used to create a new entry for location of soldier.Automatically created from data sent from sensors.
-    def __init__(self, long, lat, grid, time, soldier_id):
-        """
-        Initializes a new entry for the location of a soldier.
-
-        Parameters:
-            long (str): Longitude of the soldier's location.
-            lat (str): Latitude of the soldier's location.
-            grid_reference (str): Grid reference of the soldier's location.
-            update_time (str): Time when the location entry was made.
-            soldier_id (int): ID number of the soldier to whom the entry belongs.
-        """
-        self.long = long  # Longitude.
-        self.lat = lat  # Latitude.
-        self.grid_reference = grid  # Grid reference.
-        self.update_time = time  # Time location entry was made.
-        # ID number of soldier that the entry belongs to.
-        self.soldier_id = soldier_id
-
-
-################################################
-############ View Functions#####################
-
-@app.route('/')  # flask app displays home page
+@main.route('/')  # flask app displays home page
 def index():
     """
     Flask Route: '/'
@@ -291,11 +31,11 @@ def index():
     Returns:
     - str: Rendered HTML page ('index.html').
     """
-    return render_template('index.html')  # Returns 'home.html' to be displayed to user.
+    return render_template('    qAwindex.html')  # Returns 'home.html' to be displayed to user.
 
 
 # function to add a section
-@app.route('/add-section', methods=['GET', 'POST'])
+@main.route('/add-section', methods=['GET', 'POST'])
 def add_section():
     """
     Flask Route: '/add-section' (GET and POST)
@@ -342,7 +82,7 @@ def add_section():
 
 
 # function to add a section
-@app.route('/add-amunition', methods=['GET', 'POST'])
+@main.route('/add-amunition', methods=['GET', 'POST'])
 def add_ammo():
     """
     Flask Route: '/add-amunition' (GET and POST)
@@ -382,7 +122,7 @@ def add_ammo():
     return render_template('add-amunition.html', form=form)
 
 
-@app.route('/section-overview')  # Displays section overview to user
+@main.route('/section-overview')  # Displays section overview to user
 def section_overview():
     """
     Flask Route: '/section-overview'
@@ -398,7 +138,7 @@ def section_overview():
     return render_template('section-overview.html')
 
 
-@app.route('/delete-section', methods=['POST', 'GET'])  # Remove section
+@main.route('/delete-section', methods=['POST', 'GET'])  # Remove section
 def delete_section():
     """
     Route handler to remove a section along with its associated soldiers.
@@ -439,7 +179,7 @@ def delete_section():
 
 
 # add person to section,uses army number as primary key to add member to section. initial values automatically given.
-@app.route('/add-soldier', methods=['GET', 'POST'])
+@main.route('/add-soldier', methods=['GET', 'POST'])
 def add_soldier():
     """
     Flask Route: '/add-soldier'
@@ -504,7 +244,7 @@ def add_soldier():
 
 
 # display  list of soldiers heart rate to user
-@app.route('/list-heart-rate/<id>')
+@main.route('/list-heart-rate/<id>')
 def list_heart_rate(id):
     """
     Flask Route: '/list-heart-rate/<id>'
@@ -530,7 +270,7 @@ def list_heart_rate(id):
 
 
 # select which section members heart rate to view
-@app.route('/view-heart', methods=['POST', 'GET'])
+@main.route('/view-heart', methods=['POST', 'GET'])
 def view_heart():
     """
     Flask Route: '/view-heart'
@@ -557,7 +297,7 @@ def view_heart():
 
 
 # Select soldier to view  locations soldier has been
-@app.route('/list-location/<id>')
+@main.route('/list-location/<id>')
 def list_location(id):
     """
     Flask Route: '/list-location/<id>'
@@ -583,7 +323,7 @@ def list_location(id):
 
 
 # displays locations soldier has been
-@app.route('/view-location', methods=['POST', 'GET'])
+@main.route('/view-location', methods=['POST', 'GET'])
 def view_location():
     """
     Flask Route: '/view-location'
@@ -609,7 +349,7 @@ def view_location():
     return render_template('view-location.html', form=form)
 
 
-@app.route('/list-section-members')  # list all members in section
+@main.route('/list-section-members')  # list all members in section
 def list_section_members():
     """
     Flask Route: '/list-section-members'
@@ -626,7 +366,7 @@ def list_section_members():
 
 
 # sends section information to a page in json script format.
-@app.route('/section', methods=['GET'])
+@main.route('/section', methods=['GET'])
 def section_json_information():
     """
     Flask Route: '/section'
@@ -654,7 +394,7 @@ def section_json_information():
 
 
 # sends section information to a page in json script format.
-@app.route('/soldiers', methods=['GET'])
+@main.route('/soldiers', methods=['GET'])
 def soldier_json_information():
     """
     Flask Route: '/soldiers'
@@ -683,7 +423,7 @@ def soldier_json_information():
 
 
 # remove soldier from section
-@app.route('/delete-soldier', methods=['POST', 'GET'])
+@main.route('/delete-soldier', methods=['POST', 'GET'])
 def delete_soldier():
     """
     Flask Route: '/delete-soldier'
@@ -725,7 +465,7 @@ def delete_soldier():
 
 
 # takes information sent by arduino and updated  database
-@app.route('/sensor-data/<message>')
+@main.route('/sensor-data/<message>')
 def sensor_data(message):
     """
     Flask Route: '/sensor-data/<message>'
@@ -903,6 +643,4 @@ def sensor_data(message):
     return "ok"  # message returned to arduino by server.
 
 
-if __name__ == '__main__':  # if python is run as main file.
-    app.run(debug=True)  # turns on debugger used while testing code on laptop.
-    # app.run(host = '0.0.0.0',port = '5000')#Allows access to server when set to'0.0.0.0' uncomment to when running on raspberry pi.
+
